code duplication fix
