+++ src/components/bottomTab/AddOptionModal.tsx
@@ -78,15 +78,11 @@ const AddOptionModal = (props: Props) => {
   }
 
   const provideLocation = () => {
-    try {
       if (GPSLocation[0] === 0) {
         userCurrentLocation().catch((error) => { // Use .catch() to handle errors
           console.log("Error:", error); 
         });
       }
-    } catch (error) {
-      console.log("Error:", error);
-    }
   };
 
   const addOptions = [
