+++ src/components/previewIntervention/SampleTreePreviewList.tsx
@@ -20,6 +20,7 @@ import { useSelector } from 'react-redux'
 import { RootState } from 'src/store'
 import { nonISUCountries } from 'src/utils/constants/appConstant'
 import { INTERVENTION_STATUS } from 'src/types/type/app.type'
+import { convertMeasurements } from 'src/utils/constants/measurements'
 
 interface Props {
   sampleTress: SampleTree[]
@@ -64,37 +65,6 @@ const SampleTreePreviewList = (props: Props) => {
     navigation.navigate("TreeRemeasurement", { interventionId: interventionId, treeId: id })
   }
   const isNonISUCountry: boolean = nonISUCountries.includes(Country);
-
-
-
-  function convertMeasurements(value, unit) {
-    // Conversion factors
-    const metersToFeet = 3.28084;
-    const cmToInches = 0.393701;
-    if (isNonISUCountry) {
-        // Convert based on the unit type
-        if (unit === 'm') {
-            // Convert meters to feet
-            const feet = (value * metersToFeet).toFixed(2); // rounding to 2 decimal places
-            return `${feet}`;
-        } else if (unit === 'cm') {
-            // Convert centimeters to inches
-            const inches = (value * cmToInches).toFixed(2); // rounding to 2 decimal places
-            return `${inches}`;
-        } else {
-            return 'Invalid unit provided'; // Handle invalid units
-        }
-    } else {
-        // If ISU country, return the value as is
-        if (unit === 'm') {
-            return `${value}`;
-        } else if (unit === 'cm') {
-            return `${value}`;
-        } else {
-            return 'Invalid unit provided'; // Handle invalid units
-        }
-    }
-}
 
 
   const hasDetails = sampleTress && sampleTress.length > 0
@@ -155,7 +125,7 @@ const SampleTreePreviewList = (props: Props) => {
               <Text style={styles.iconTitle}>{i18next.t("label.height")}</Text>
               <View style={styles.iconMetaWrapper}>
                 <HeightIcon width={10} height={20} />
-                <Text style={styles.iconLabel}> {convertMeasurements(details.specie_height, 'm')} {isNonISUCountry ? i18next.t('label.select_species_feet') : 'm'}</Text>
+                <Text style={styles.iconLabel}> {convertMeasurements(details.specie_height, 'm', isNonISUCountry)} {isNonISUCountry ? i18next.t('label.select_species_feet') : 'm'}</Text>
               </View>
             </View>
             <View style={styles.iconWrapper}>
@@ -164,7 +134,7 @@ const SampleTreePreviewList = (props: Props) => {
                 <View style={styles.iconHolder}>
                   <WidthIcon width={20} height={20} />
                 </View>
-                <Text style={styles.iconLabel}> {convertMeasurements(details.specie_diameter,'cm')} {isNonISUCountry ? i18next.t('label.select_species_inches') : 'cm'}</Text>
+                <Text style={styles.iconLabel}> {convertMeasurements(details.specie_diameter,'cm', isNonISUCountry)} {isNonISUCountry ? i18next.t('label.select_species_inches') : 'cm'}</Text>
               </View>
             </View>
           </View>
